- feat: Add ApiClient services as companion object
This commit adds ApiClient services as a companion object. By defining the ApiClient services as a
companion object, developers can easily access and use the services without the need to create a new
instance of the ApiClient class. This change will make the code more efficient and easier to read.