- style: update Max Button UI based on inputted and user balance
This commit updates the UI of the Max Button to reflect changes in the inputted amount and user
balance. If the inputted amount is equal to the user balance, the Max Button UI is modified to
indicate that the maximum amount has been selected. This enhancement improves the user experience by
providing a clear visual indication of the selected amount.